fixing negative event numbers. Details: https://its.cern.ch/jira/browse/ATLASVPONE-287. (VP1Base-00-23-12)
    * fixing negative event numbers.
	  Details: https://its.cern.ch/jira/browse/ATLASVPONE-287.
	* new tag: VP1Base-00-23-12


Former-commit-id: 41f7ab6c41ff1034bab631977fb234c533bec00e

+++ graphics/VP1/VP1Base/VP1Base/IVP1ChannelWidget.h
@@ -48,7 +48,7 @@ public:
 
   virtual bool isAccumulator() const;
 
-  void setRunEvtNumber(int runnumber, int eventnumber);
+  void setRunEvtNumber(int runnumber, unsigned long long eventnumber);
 
 protected:
   virtual void create();//Fixme: Abstract!
@@ -65,7 +65,7 @@ protected:
   void turnOn(IVP1System*);
   void turnOff(IVP1System*,const bool& immediateErase=true);
 
-  void getRunEvtNumber(int& runnumber, int& eventnumber);
+  void getRunEvtNumber(int& runnumber, unsigned long long& eventnumber);
 
   void setEvtTimestamp(unsigned timestamp);
   void getEvtTimestamp(unsigned& timestamp);
